refactor: move of files inside src/components
refs: FILES-645
