+++ src/components/BadgeWrap.tsx
@@ -3,7 +3,12 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import React, { FC, forwardRef } from 'react';
+/*
+ * SPDX-FileCopyrightText: 2023 Zextras <https://www.zextras.com>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+import React from 'react';
 
 import { Container, Icon } from '@zextras/carbonio-design-system';
 import { BadgeInfo } from '@zextras/carbonio-shell-ui';
@@ -20,14 +25,14 @@ const MiniBadge = styled(Icon)`
 	z-index: 99;
 `;
 
-// eslint-disable-next-line react/display-name
-const BadgeWrap: FC<{ badge: BadgeInfo }> = forwardRef<HTMLDivElement, { badge: BadgeInfo }>(
-	({ badge, children }, ref) => (
+export const BadgeWrap = React.forwardRef<
+	HTMLDivElement,
+	{ badge: BadgeInfo; children: React.ReactNode }
+>(function BadgeWrapFn({ badge, children }, ref) {
+	return (
 		<Container width={'3rem'} height={'3rem'} style={{ position: 'relative' }} ref={ref}>
 			{badge.show && <MiniBadge color={badge.color} icon="AlertCircle" />}
 			{children}
 		</Container>
-	)
-);
-
-export default BadgeWrap;
+	);
+});
